Manage repository using settings file
The configuration for this repository is now managed using the
repository-settings[^1] app. The app reads the `.github/settings.yml`
file in this repository and applies the settings in the file to the
repository. This allows us to fully manage the repository in code,
which is especially handy when modifying the GitHub Actions workflows.

[^1]: https://github.com/repository-settings/app