Set up crates for crux's example
We are following the getting started guide[^1] of crux to get the
development environment set up. The first part of the guide sets up
various crates, which has been done. The second part, which sets up an
iOS project, will follow in another commit.

Additionally, we intend to remove most of the example code once we have
it working. At that point, we will also set up proper tooling to format,
lint, and test the app.

[^1]: https://redbadger.github.io/crux/getting_started/core.html