Set up iOS app (#25)
We are following the getting started guide[^1] of crux to get the
development environment set up. The first part set up the Rust crates,
while the second part focused on the iOS app. Getting the app to build
required making some changes to the crates, but the project builds
successfully now and can be run in an iOS simulator.

[^1]: https://redbadger.github.io/crux/getting_started/iOS/with_xcodegen.html